refactor: reorder nut18
